Refactor models Broker and Profile
