--- Common/src/models/containers/broker.js
@@ -1,8 +1,0 @@
-class Broker {
-    constructor(options = {}) {
-        this.host = options["host"] || "localhost"
-        this.port = options["port"] || 1883
-    }
-}
-
-module.exports = Broker
