+++ Common/src/utils/database/mongodb.js
@@ -1,14 +1,14 @@
 const mongoose = require("mongoose")
 const logger = require("../logging/logger").logger
 
-function initialize(mongo) {
-    mongoose.connect(mongo.getURIString(), mongo.getOptions()).then(
-        () => logger.info("Mongo successfully connected"),
-        err => {
+async function initialize(mongoDb) {
+    await mongoose.connect(mongoDb.getURIString(), mongoDb.getOptions(), err => {
+        if (err) {
             logger.error(err)
             process.exit()
         }
-    )
+        logger.info("MongoDB successfully connected")
+    })
     return mongoose
 }
