+++ Device/src/backend/profile/profile.js
@@ -1,8 +1,15 @@
-const Profile = require("common").models.containers.Profile
-
-class DeviceProfile extends Profile {
+class Broker {
     constructor(options = {}) {
-        super(options)
+        this.host = options["host"] || "localhost"
+        this.port = options["port"] || 1883
+    }
+}
+
+class DeviceProfile {
+    constructor(options = {}) {
+        this.backendId = options["backendId"] || "group3-be"
+        this.broker = new Broker(options["broker"] || {})
+        this.active = options["active"] || true
         this.device = options["device"]
     }
 }
